Update clouds.yaml to follow a profile authorization convention
It's handy to have the authorization for a profile baked into the name
so you can do things like this:

  $ openstack user list --os-cloud devstack-system-admin
  $ openstack server list --os-cloud devstack-demo-reader
  $ openstack volume create --os-cloud devstack-alt-member

Where the middle portion of the name is the authorization target (e.g.,
project name or system) and the last bit is the role (e.g., admin,
member, or reader).

With more OpenStack projects adopting secure RBAC personas, this makes
it easier to tinker with different situations locally if you use
clouds.yaml.

Change-Id: Ide5c475a434ade01b25b111e35bbfe0c702a1ce8