Update patch set 3
Patch Set 3:

> this is not needed as discussed in https://review.opendev.org/c/openstack/tempest/+/779814

Note that it's not needed only if we go the admin-only-for-create route. There are other issues with this API besides just unbounded create, so I'll leave this open in case we decide to keep the stricter approach. But yes, if we do the admin thing, then this goes away.

Patch-set: 3