Use specific credentials for tempest
Tempest expects the classic environment variables to be
present for credentials to access the cloud, but this is
wrong in cases where we're trying to setup system
scoped services and need to remove the environment variables
that was being used.

Instead, change the plugin to use the os-cloud entry definitions,
and specifically in this case devstack-admin which makes sense
until we begin to start to make tempest scope aware.

Change-Id: I8d4ec68f116eea07bc7346f939e134fa2e655eac