Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> (1 comment)
> 
> See https://review.opendev.org/c/openstack/devstack/+/780417 for a different idea that ensures that using the cloud config actually works and also avoids having to specify the cloud for every single command.

So I *suspect* that in many places it will still need to be defined, and may even be good for people to accidentally try and change those around, and eventually will even need to be configurable. I guess a good example of it being a positive thing for the repetition is that we didn't know glance would accept but not make accessible an image uploaded by system-admin until ?Yesterday? (Or maybe it was Thursday.)

Patch-set: 2