Add route to IPv6 private subnets in ML2/OVN
Reproduce the behavior from ML2/OVS: add a route from the devstack node
to the CIDRs of the default IPv6 subnet pool. Any IPv6 subnet created
using the default subnet pool and plugged into the default router is
reachable from the host (ipv6-private-subnet).

The commit also removes the IPV6_ROUTER_GW_IP global variable, it was
unused and it created some confusion with the real IP address of the
default router.

Change-Id: I02ca1d94e9f4d5ad4a06182f5ac9a2434941cf08