Ensure that /etc/openvswitch/system-id.conf file exists
In the neutron's ovn-agent module devstack tries to save generated, or
set by the user system-id to the /etc/openvswitch/system-id.conf file.
It was added in [1]. But it seems that during stack.sh on the fresh
node this file may not exists so we should ensure it exists always.

[1] https://review.opendev.org/c/openstack/devstack/+/779991

Related-Bug: #1918656

Change-Id: If4cb83c4945cb91df1ba7fb89e29a1d0b915a28a