Use (or set properly) system-id generated by openvswitch (cont.)
This is a continuation of commit 1ed276c17791dba1f0b7ef4446d0efe09135553b

When setting the system-id generated by openvswitch, it is possible that
the $OVS_SYSCONFDIR is not yet created. That is expected to happen when
building OVN from source (OVN_BUILD_FROM_SOURCE=True). In such cases,
it we just need to create that directory before setting system-id.conf.

Change-Id: I68306a50cba7dd8d47622f2dd07ec4c41c544ede