Adapt `ceph osd pool set size` command for pacific
As of the Ceph Pacific release, by default setting a pool's size
to one replica requires an additional 'yes-i-really-mean-it'
argument [1.

Add this argument where cinder backup pool size is set so that
it will not fail [2] when run with pacific or newer.  The extra
argument does no harm with older Ceph branches and releases.

[1] https://ceph.io/releases/v16-2-0-pacific-released/
[2] https://zuul.opendev.org/t/openstack/build/bb2a76faa1c647759c413afba52e6b7c/log/controller/logs/devstacklog.txt#7119

Change-Id: I42940573cf8be9e3a93a225ecee9066e3969170a