Update patch set 1
Patch Set 1: Code-Review+1

So this is actually just moving all of the tasks from the role into a block with 'failed_when: False' to ignore any error. Due to this is a post task and we don't want to fail if there's any error during the run of the role, this looks good!

Thanks for the patch Ghanshyam! If I'm not mistaken we can backport this all the way till pike and fix the gate there!

Patch-set: 1
Reviewer: Gerrit User 17685 <17685@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1