Update patch set 1
Patch Set 1:

Just for the record, neutron-grenade seems broken [1] in stable/train (and only in stable/train), because in the 'smoke test' run it tries to use the upper constrain from /opt/stack/requirements/upper-constraints.txt, but it does not exist. I guess because it's either under /opt/stack/old/requirements/ or /opt/stack/new/requirements. I still try to figure out how is it exactly. (Maybe related to this patch? [2])

[1] https://zuul.opendev.org/t/openstack/build/9e521168cb264eba8e4af36e4ba383d0/log/logs/grenade.sh.txt#28855
[2] https://review.opendev.org/q/I66b5d128c3dbe51289410760ea1cbf7f61c55c2f

Patch-set: 1
CC: Gerrit User 17685 <17685@4a232e18-c5a9-48ee-94c0-e04e7cca6543>