Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> in failing neutron-grenade job also constraint var are correct path
> 
> https://zuul.opendev.org/t/openstack/build/9e521168cb264eba8e4af36e4ba383d0/log/logs/grenade.sh.txt#28560
> 
> TEMPEST_VENV_UPPER_CONSTRAINTS=/opt/stack/old/requirements/upper-constraints.txt
> - https://zuul.opendev.org/t/openstack/build/9e521168cb264eba8e4af36e4ba383d0/log/logs/grenade.sh.txt#438
> 
> But not sure how different value is picked while running test
> - https://zuul.opendev.org/t/openstack/build/9e521168cb264eba8e4af36e4ba383d0/log/logs/grenade.sh.txt#28850

I couldn't figure out why it changes as there is no trace of it in the logs :-/ Nevertheless, it seems neutron-grenade works in stable/train of nova repository, there only the nova-grenade is failing. So it's not really consistent.

Anyway, with the above linked workaround train gate could be fixed (see:

https://zuul.opendev.org/t/openstack/build/c3ab14800d404e898b8b9848b0eb2d58/log/logs/grenade.sh.txt#40035-40036 )

I'll propose the same patch on master, as I guess this will be an issue in ussuri, when train will transition to Extended Maintenance, so let's fix it in advance.

Patch-set: 1