Add some debug to async_wait failures
This dumps some data in the case where we fail to wait for a child
pid to help debug what is going on.

This also includes a change (which needs to be cleaned up) which
holds the children hostage when they have completed their tasks,
until signed by the wait process to avoid them sitting in completed
state for a long period of time. This may work around an apparent
bug in bash process accounting.

Change-Id: I7b58ce0cf2b41bdffa448973edb4c992fe5f730c
Related-Bug: #1923728