Update patch set 2
Patch Set 2: Code-Review+2

if this solves the test failures, I agree, 6GB is not that huge increase and it's just a default value/recommended one because it's proven it works as it's tested - that of course doesn't mean that the things don't work with less memory. The value can be still overridden if it's needed for a local environment.

Patch-set: 2
Reviewer: Gerrit User 22873 <22873@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2