Update patch set 3
Patch Set 3:

Sorry, yes, I forgot about this patch, maybe because of the merged workaround. The problem is that I don't know that this would work in every cases. Most of the cases, what I've tested, it worked. But there could be cases where something is (force?-)reinstalled and that could pull in wrong versions from pypi. :/

If you think that this could be merged, and we will see (and fix) later such failing cases, then let me know and I'll try to get back to this patch.

Patch-set: 3