Update patch set 1
Patch Set 1:

Be careful doing this. If the distro pbr package is built with distutils and not setuptools then pip won't be able to clean it up later. (Whether or not that is an issue I don't know)

Another option is to simply preinstall pbr with pip. Pip doesn't have the SNI issues that old easy_install has.

Patch-set: 1
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>