Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2 Review-Priority+2
> 
> thanks Ian. not sure why it did not occur in https://review.opendev.org/c/openstack/devstack/+/735097

I believe the reason it works in normal gating is that we "preclone" and configure all of the git repos before devstack runs. This ensures that depends-on works. But a side effect is that you can refer to code that won't be present if someone has cloned devstack and run stack.sh.

Patch-set: 2
CC: Gerrit User 4146 <4146@4a232e18-c5a9-48ee-94c0-e04e7cca6543>