Update default enabled services to handle OVN default
Neutron recently switched to using OVN by default in Devstack but didn't
update the default enabled services list in stackrc. This results in
errors like:

  [ERROR] /opt/devstack/lib/neutron_plugins/ovn_agent:360 The q-agt/neutron-agt service must be disabled with OVN.

We update the default enabled services here to address this.

Note that this was missed in CI because the CI job was modified to set
the enabled services. The CI jobs also set a few other env vars that we
may need to set by default in stackrc but one thing at a time.

Change-Id: Ifa41cc3e651e5afb76a468bdef705deb5f1d61ad