Update patch set 1
Patch Set 1:

> will this cycle ever end? 😉

I don't think it's fair to break other gates.  I'm trying to setup something to at least debug this, and https://review.opendev.org/c/zuul/nodepool/+/793273 should hopefully test this revert to pinpoint this as the source of the problem, and also hold a node in a good state so we can compare.

If this is some sort of minor config thing that would be great.  But we also need to unblock nodepool and dib gates, at least.

Patch-set: 1