Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I only don't understand why it gets sourced always now, instead of only when the from-source build is requested.

I thought it was more "tidy" to have that include together with the rest of the includes. Also, it was like that before we moved it to inside that OVN_BUILD_FROM_SOURCE == True check to make the "ensure-devstack" job happy.

But, if reviewers think it's best to keep it inside the conditional, I've no problem with that either.

Patch-set: 3
Reviewer: Gerrit User 6773 <6773@4a232e18-c5a9-48ee-94c0-e04e7cca6543>