[OVN] PUBLIC_NETWORK_GATEWAY defaults to public IPv4 subnet GW
If FLOATING_RANGE and OVN_L3_CREATE_PUBLIC_NETWORK are set, a new
subnet is created in the external network.

The external gateway used in this subnet will be set in the public
bridge external interface. If PUBLIC_NETWORK_GATEWAY is not defined,
a default value will be assigned by Neutron.

In case of IPv6 networks, IPV6_PUBLIC_NETWORK_GATEWAY must be defined.

Closes-Bug: #1930360

Change-Id: Ic177ce4f3accb462607084f17e317fee3f94681a