Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > You might consider keeping the old default of 1GB then override that in your jobs to 5GB with the new var. But I expect 5GB is safe enough.
> 
> Yeah, I mirrored those off the default limits for things like cinder, which are based on disk image default sizes for things like ceph. Those also weren't tunables, so I didn't make this one either.
> 
> I'd say leave the default at 1G and override to whatever you need in a job, especially with it being a variable, personally.

I can see the point of view that 1GB should be sane and fine, but the fact the addition of the limit has broken something means it is going to be a breaking for others, which is just not great experience wise for someone trying to figure things out and experiment unless they already also happen to know and understand the limits interface, behaviors, and already have the in-depth openstack development knowledge to hunt down what exactly is going on.

I guess I'm also conflicted. If we were a solely ubuntu-centric community, we wouldn't likely have this issue. The Ubuntu images are ~600 MB, where as the Centos Stream images are coming in a 1.3GB at present. Which means, any developer is going to hit a non-specific gateway error when trying to upload a stock centos stream image to glance. I guess it may be delineated if we've turned it on globally for all zuul jobs and it is opt-in only for manual devstack runs, but I didn't dig that far as we can't merge patches in ironic until we've got this sorted.

Patch-set: 1