Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Let's first check the behaviour in CI via https://review.opendev.org/c/openstack/devstack/+/795640
> 
> I think the logic we have here should either go away for all the distros or stay. Otherwise, we are marching straight into chaos.
> 
> Also, we should probably remove older fedoras from the support list.

Sure, good to know that. IMO, 795640 will finnaly meet the errors which are fixed by this patch. For now, that patch does not work at all.

If pip is built-in for all python3 distros, these removal here can be applied to all OS forever. It should be fine.

Patch-set: 1