Update patch set 1
Patch Set 1:

> Patch Set 1:
> Slightly off-topic but what has changed w.r.t. image building in F34?

We used to build Fedora with our "minimal" element which sets up a chroot and uses rpm on the host system to extract the base system.  Fedora 34 changed the compression format for RPM's and Debian's rpm (the version used in the container we use to build) didn't understand it and thus couldn't extract things.

We've moved to "containerfile" based Fedora 34 builds, which takes the upstream docker container and extracts that, effectively giving us the base chroot to start from.  It blurs the line a bit between what is dib and what is done in a Dockerfile; but the nice thing is all work done to build all the existing elements "just work".

Really, this is like using the upstream .iso's which we did a long time ago.  But we found they a) constantly got bigger and bigger with more stuff we didn't need and b) liked to change things rather randomly, because they weren't meant to be built ontop of.  docker images largely solve these pain points so it feels like a much better fit.

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>