Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > thanks, can you backport this to wallaby also so that we can verify the installation of stable branch block.
> > > > 
> > > > We do not have any stable branch job running in devstack master gate.
> > > 
> > > Yes this needs to backported in other stable releases as well atleast till ussuri. I see you already backported https://review.opendev.org/c/openstack/devstack/+/803071, Thanks. In noticed c8-centos job running there, RDO wallaby is c8-stream only so would need to switch c8 job to c8-stream
> > 
> > humm this is good question :) if I am understanding correctly, you are saying it would not work for wallaby but can work on other older stable?
> 
> For wallaby need to use c8-strem job devstack-platform-centos-8-stream as RDO wallaby released only for c8-stream so may not work for c8, /me can work on adding that in wallaby, Ussuri and Victoria should work with both c8 and c8-stream.

humm, Actually thing is our testing runtime for wallaby is c8 not c8-stream.  we need to discuss it in TC and QA if we can update the the runtime at this time or not or just remove c8 from wallaby gate as it cannot be fixed by c8 ?

Let me bring this to TC/QA meeting first but this fix is all good backport and how to make centos job work in wallaby is separate discussion. Thanks Yatin for all the updates and fixes.

Patch-set: 1