Update patch set 1
Patch Set 1:

Note about the centos-8 job failure: devstack does build on centos-8-stream in wallaby (see [0] for an example).  The centos-8 job is failing right here, but that's due to issues independent of this patch.  This patch works in stable/victoria where devstack is building on centos-8 (see [1] for an example).

[0] https://review.opendev.org/c/openstack/python-cinderclient/+/799882, the python-cinderclient-functional-py36 job in zuul check
[1] https://review.opendev.org/c/openstack/python-cinderclient/+/800635, also the python-cinderclient-functional-py36 job

Patch-set: 1