Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Note about the centos-8 job failure: devstack does build on centos-8-stream in wallaby (see [0] for an example).  The centos-8 job is failing right here, but that's due to issues independent of this patch.  This patch works in stable/victoria where devstack is building on centos-8 (see [1] for an example).
> 
> [0] https://review.opendev.org/c/openstack/python-cinderclient/+/799882, the python-cinderclient-functional-py36 job in zuul check
> [1] https://review.opendev.org/c/openstack/python-cinderclient/+/800635, also the python-cinderclient-functional-py36 job

yeah, in wallaby centos8 was testing runtime so we need to see if we can change that to centos8-stream at this time or not? I am going to discuss this in QA and TC meeting this week and let's see. IMO, changing the testing runtime for wallaby is not good but it seems there is no way to continue testing the centos8 on wallaby?

Patch-set: 1
Reviewer: Gerrit User 8556 <8556@4a232e18-c5a9-48ee-94c0-e04e7cca6543>