Add option to set chap algoorithms for iscsid for FIPS
The default CHAP algorithm for iscsid is md5, which is disallowed
under fips.  We need an option to set the chap algorithms to
something other than md5 to allow fips testing on devstack nodes.

The new environment paramter is ISCSID_ENABLE_FIPS.  When this
is "True", then the iscsid chap algorithms will be set to
"SHA3-256,SHA256", which works under FIPS.

Change-Id: Ide186fb53b3f9826ff602cb7fb797f245a15033a