Update patch set 1
Abandoned

If I remember correctly there was a 'catch 22' around stestr version in stable/ocata, so this patch is not enough to make the gate working. Besides the gate of devstack's stable/ocata has been hanging around unfixed for quite some time now, so it's time to announce End of Life for it due to lack of maintenance.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon