Update patch set 4
Patch Set 4:

I guess that since this is an overall security thing for iscsid putting it in stack.sh is appropriate.  

Maybe it would be more useful to have a flag ISCSID_ENABLE_FIPS or something that selects appropriate algorithms; so that we are encoding what is expected to work in here (rather that random tests specifying their own, in different ways, etc)?

Patch-set: 4
CC: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>