Add a common function for system role assignments
As more and more services adopt system scope across OpenStack we're
going to need a function to grant system role assignments to users,
including service users.

This commit introduces a new function that creates system role
assignments in keystone and it's similar to the existing functions we
have for creating project role assignments.

Needed-By: https://review.opendev.org/c/openstack/tempest/+/805452
Change-Id: Ie26b5770304bc43b9bde36efe1eda29a520fc287