Ensure the nova service user is a system-admin
Neutron calls some system-specific nova APIs/policies
(os_compute_api:os-server-external-events:create). When nova is
configured to enforce scope, these calls fail because the nova service
doesn't have the admin role on the system.

We are expecting to hit more cases like this as more services adopt
system scope. This commit uses a common function to grant the nova
service user the admin role on the system after it's created.

Needed-By: https://review.opendev.org/c/openstack/tempest/+/805452
Change-Id: I000209a4677d7db5622016667ea7d38626d50918