Revert "Replace dstat with pcp-dstat"
This reverts commit c19c1262c8b81f1cc543eafb9e5c3a34c4b632fe.

The pcp package fails to install on ubuntu fairly regularly due to a
timeout starting the pmlogger.service unit. It seems there have been
bugs for this in fedora [0] and while that bug notes a fix was
upstreamed that fix does not seem to be sufficient.

You can see the occurence of these failures by querying logstash for:

  message:"pmlogger.service failed because a timeout was exceeded." AND filename:"job-output.txt"

I'm pushing this change to start some conversation about the problem. If
we don't want to revert we should probably add a devstack bug for this
issue and an e-r entry to track it.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1721223

Change-Id: I657e802805db6835693e6b811799b189a71bec0d