Update patch set 4
Patch Set 4:

Based on the fact we need EPEL for this and it's not really maintained anyway - I suggest we stay away from it as well. PCP is an overkill. Perhaps let's just go with some custom script I'll make up to debug Debian. I also don't think this is useful for dev installation of devstack. Most likely this was fudged in to get better insight in CI. I'll add it to QA meeting agenda.

Patch-set: 4