Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> recheck we haven't seen this ceph issue before, lets assume it is something transitional

FYI we have seen this before [1][2] and it means the ceph version being installed is >= pacific, they made a change that disallows setting a pool size of 1 without passing the “yes-I-really-mean-it” flag.

Looks like the ceph version for tempest-py3 jobs was recently updated to pacific [3] but I’m not sure whether that’s related. Devstack on stable/victoria is trying to configure a pool size of 1 [4]. I don’t know how/why a job on stable/victoria is using pacific as the change in devstack-plugin-ceph was done on only the master branch. Maybe it’s something to do with the nova-ceph-multistore job specifically?

[1] https://review.opendev.org/c/openstack/nova/+/785059/3#message-c31738db1240ddaa629a3aaa4e901c5a62206e85
[2] https://meetings.opendev.org/irclogs/%23openstack-nova/%23openstack-nova.2021-05-27.log.html#t2021-05-27T15:36:28
[3] https://github.com/openstack/devstack-plugin-ceph/commit/ea0c1c768c99437504a1a6f2971dd0c5c1563ca5
[4] https://github.com/openstack/devstack/blob/stable/victoria/lib/cinder_backends/ceph

Patch-set: 1
CC: Gerrit User 4690 <4690@4a232e18-c5a9-48ee-94c0-e04e7cca6543>