Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> (1 comment)
> 
> This could break existing initiators on restacking, no? I suggest we either do this for infra itself only or make devstack track whether it has already set this (using a helper file for example). Share your thoughts.

As I said in the commit, I can't imagine anyone is using an existing initiator within a devstack env before stacking or restacking. Within the context of the deployment any prior use of the initiator is obviously removed during a restack so that also shouldn't be an issue IMHO.

If it later transpires that we do have users who need to retain the original initiator name then we can always add a variable to control when this takes place, defaulting to True for the main use case that they can disable.

Patch-set: 5