Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> nova-gerande-multinode is passing in https://review.opendev.org/c/openstack/nova/+/812095 but stable/stein (this patch) hitting some novnc issue

I looked at the novnc fail for some time and I can see what is wrong, this "Path:" is wrong [1]:

 INFO nova.console.websocketproxy [-] 10.209.37.246 - - [01/Oct/2021 02:03:14] 10.209.37.246: Path: '/websockify'

this will cause us to not be able to parse out the token in the console proxy. The token is supposed to be in the path, like this [2]:

 INFO nova.console.websocketproxy [-] 10.0.1.216 - - [30/Sep/2021 03:10:15] 10.0.1.216: Path: '/?token=f8a8bc2e-1985-43bc-9523-3efe81b2d043'

I think I finally found why, I think it's because in tempest 20.0.0 we don't have the necessary commits for handling the "path" parameter [3][4]. These commits were introduced in tempest 24.0.0.

[1] https://zuul.opendev.org/t/openstack/build/e593cef5e83f4c308b138eee86bcd83d/log/controller/logs/screen-n-novnc-cell1.txt#13
[2] https://zuul.opendev.org/t/openstack/build/fdb42f68ef2a47c5a393d27294b16626/log/controller/logs/screen-n-novnc-cell1.txt#11
[3] https://github.com/openstack/tempest/commit/27ba933ab5db4dc50e0b557387f9e62575602756#diff-293717cdbe22a4182a68ab67a68cd0c7cccb5ac25502bf772bbda8e733d54bc2R395
[4] https://github.com/openstack/tempest/commit/ca0fad075d80ab8b8bc9997656bbd34a3f8d37a8#diff-293717cdbe22a4182a68ab67a68cd0c7cccb5ac25502bf772bbda8e733d54bc2R420

Patch-set: 1