add optional qemu wrapper
upstream qemu has recently increased the tb_size of instances
form 32MiB to 1GiB.
https://github.com/qemu/qemu/commit/600e17b261555c56a048781b8dd5ba3985650013

This is not currently tunable via libvirt and as a result cannot
be set via nova. While the libvirt feature is being implemented
this change introduces a qemu wrapper that will rewrite the qemu
command line such that the tb_size can be clampped to resonsable size
for our upstream ci.

This behavior can be enabled by setting USE_QEMU_WRAPPER=True.
By default the qemu wrapper is not enabled.

When used the tb_size will be limited to 32MiB but that can be
altered by defining TB_SIZE_LIMIT to an alternitive value.
32MiB is the old default and should work in most cases.

This change enables the qemu wrapper by default in zuul jobs.

Change-Id: I8b9cde70b5909913b7d8785e2fc398e64f93e640