Create data dir by regular user for OVN built from source.
With recent changes, there is an assumption, that all the data should be
created/removed using elevated permissions, otherwise there would be
leftovers. This change however make impossible to use OVN built from
source, due to the fact, that data dir should be accessible by regular
user. It's being fixed with this patch.

Closes-Bug: #1952467
Change-Id: If1feb4d6b54c0e57ed1d5faedc028ca6eb35019f