Allow skip the database server installation
This patch allows to skip the install of the MySQL
database package with the introduction of the
INSTALL_DATABASE_SERVER variable (defaulted to True).
This is useful in such environments that do not require
to install the MySQL server packages directly but using
a container serving that purpose, for those cases all the
remaining steps should be executed just skipping the
packages install.

Change-Id: I26628a31fdda3ce95ed04a2b7ae7b132c288581f