set default LIBVIRT_CPU_MODEL based on ARCH
This change mainntains the current default for x86_64
updates LIBVIRT_CPU_MODEL to max on aarch64 and
qemu64 for lack of a better default on anything else.

To ensure correct defaults, fallback to qemu is moved to
stackrc and aarch64 host are only set to host-passthrough
if kvm is available.

Change-Id: Ie3175069db75c4b3a7a3cfd9df136017c22618df