remove egg-link files on unstack
on some distros (ubuntu 20.04 aarch64) pip will refuse to uninstall
packages that are not located under /usr.
This breaks unstacking and restacking as the second install can
refuse to uninstall the exising copy.

to adress this in addtion to removing the pyc files unstack.sh will
now remove the egg-link files so that pip no longer considers the
repos installed in $DEST (/opt/stack) as installed.

Change-Id: I6795cdf23985a2de645f512cecc57a40a4b253f1