chown checked-out repos (to be used for install) to root ownership
Per the inline note, we can't install repos unless they are owned by
root now.

Editing /opt/stack/requirements makes this a pain, we have to swizzle
around that.

Tempest clones itself using git_clone, but then operates as a user.
So create a wrapper that doesn't update the permissions.

For the record, this is all a complete mess and I do not endorse any
of it.

Change-Id: I458d32c1fc7be41ba66ebf14cca80ba796027ac0