Separate sdist building from installation
Because Git (and so PBR) should not be called on a checkout which
isn't owned by the same user, e.g. cloning as a non-root user and
then doing `sudo pip install .`, explicitly build an sdist from the
source tree first as stack and then only elevate privs when
installing the resulting tarball (with PBR's metadata already
inside) rather than relying on pip's implicit building feature.

Change-Id: Ie72f4ff8d7ff3ec6263db6ab7281ae4ed1846f53