Separate wheel building from installation
Because Git (and so PBR) should not be called on a checkout which
isn't owned by the same user, e.g. cloning as a non-root user and
then doing `sudo pip install .`, explicitly build a wheel from the
source tree first as stack and then only elevate privs when
installing the resulting package (with PBR's metadata already
inside) rather than relying on pip's implicit build callout.

Drop editable/develop installation, as we can't support it if
separating the build and install steps. The future of editable mode
installs in SetupTools is up in the air with PEP-660, and it's also
just a fundamentally broken idea to be installing packages as one
user (root) which point to files writeable by a completely different
user (stack). If we want proper editable installs, we should put
work into venv-based installs so we can finally get rid of the `sudo
pip install` antipattern which the Python packaging community and
distros alike are desperately trying to kill off.

Change-Id: Ie72f4ff8d7ff3ec6263db6ab7281ae4ed1846f53