Pin Tempest to 26.1.0 tag for stable/ussuri testing
Stable ussuri is in Extended maintenance state[1] and
we need to pin tempest in stable/ussuri testing.

We use Tempest 26.1.0 because of oslo.utils version bump in Tempest
27.0.0 and 28.0.0 which is not compatible with stable/ussuri
upper constraints so these versions do not work in stable/ussuri.
Tempest 26.1.0 version is compatible for ussuri testing from tempest
side as well as from upper constraints side.

Depends-On: https://review.opendev.org/c/openstack/tempest/+/843045

Change-Id: I5a637bdcdf61e71c7e948187f686de40f5ef0dca