Fix doc and user create script to set homedir permissions
RHEL based distros set homedir permissions to 700,
which results into failures as defined in the below bug.

Since in doc we add useradd command, it's good to
add instructions to fix the permissions there itself
instead of getting failures during installation and then
fixing it.

Also update user create script to fix permissions
to 755 if 700 is set.

Closes-Bug: #1966858
Change-Id: Id2787886433281238eb95ee11a75eddeef514293