Attempt to warn about perf regressions
This is really a rocky road with many sinkholes along the way, which
may not make it very useful.

This attempts to store static "known good" performance profiles for
given jobs, by project, and compare each run against them. If a
given metric changes by some amount over a threshold, we can warn
about it potentially being something to look at.

This provides a tool called "get_perfdata.py" which will extract
and refresh the data for a given project based on a gerrit change.
It also provides a "grok_performance.py" tool which is run in post
that does the comparison (really, against any two performance.json
files).

Change-Id: I1b02c91f9c708f789625ed9b7d958e4fc96ff622